backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
pydantic>=2.0.0
python-dotenv==1.0.0
